pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "video_gen_agent"
version = "0.1.0"
description = "AI Video Generation Agent using Google ADK"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-adk>=0.3.0",
    "google-genai>=1.0.0",
    "requests>=2.31.0",
    "moviepy>=1.0.3",
    "imageio[ffmpeg]",
    "aiosqlite>=0.19.0",
    "aiohttp>=3.9.0",
    "aiofiles>=23.0.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
video-gen = "video_gen_agent.main:main"
